[svn-r8783] Purpose:     Code cleanup
Description:
    Clean up almost all warnings from Windows builds.

Platforms tested:
    FreeBSD 4.10 (sleipnir) w/parallel
    Too minor to require h5committest

+++ c++/test/th5s.cpp
@@ -77,7 +77,7 @@ struct space4_struct {
     unsigned u;
     float f;
     char c2;
- } space4_data={'v',987123,-3.14,'g'}; /* Test data for 4th dataspace */
+ } space4_data={'v',987123,(float)-3.14,'g'}; /* Test data for 4th dataspace */
 
 /*-------------------------------------------------------------------------
  *
