+++ src/H5Dio.c
@@ -1151,7 +1151,7 @@ H5D_contig_read(hsize_t nelmts, H5D_t *dataset,
             target_size = max_type_size;
         /* If the buffer is too large to hold all the elements, make it smaller */
         else if(target_size>(nelmts*max_type_size))
-            target_size=(nelmts*max_type_size);
+            target_size=(size_t)(nelmts*max_type_size);
     } /* end if */
     request_nelmts = target_size / max_type_size;
     
@@ -1202,7 +1202,7 @@ H5D_contig_read(hsize_t nelmts, H5D_t *dataset,
     for (smine_start=0; smine_start<nelmts; smine_start+=smine_nelmts) {
         /* Go figure out how many elements to read from the file */
         assert(H5S_SELECT_ITER_NELMTS(&file_iter)==(nelmts-smine_start));
-        smine_nelmts = MIN(request_nelmts, (nelmts-smine_start));
+        smine_nelmts = (size_t)MIN(request_nelmts, (nelmts-smine_start));
 	
         /*
          * Gather the data from disk into the data type conversion
@@ -1389,7 +1389,7 @@ H5D_contig_write(hsize_t nelmts, H5D_t *dataset,
             target_size = max_type_size;
         /* If the buffer is too large to hold all the elements, make it smaller */
         else if(target_size>(nelmts*max_type_size))
-            target_size=(nelmts*max_type_size);
+            target_size=(size_t)(nelmts*max_type_size);
     } /* end if */
     request_nelmts = target_size / max_type_size;
 
@@ -1444,7 +1444,7 @@ H5D_contig_write(hsize_t nelmts, H5D_t *dataset,
     for (smine_start=0; smine_start<nelmts; smine_start+=smine_nelmts) {
         /* Go figure out how many elements to read from the file */
         assert(H5S_SELECT_ITER_NELMTS(&file_iter)==(nelmts-smine_start));
-        smine_nelmts = MIN(request_nelmts, (nelmts-smine_start));
+        smine_nelmts = (size_t)MIN(request_nelmts, (nelmts-smine_start));
 	
         /*
          * Gather data from application buffer into the data type conversion
@@ -1658,7 +1658,7 @@ H5D_chunk_read(hsize_t nelmts, H5D_t *dataset,
             target_size = max_type_size;
         /* If the buffer is too large to hold all the elements, make it smaller */
         else if(target_size>(nelmts*max_type_size))
-            target_size=(nelmts*max_type_size);
+            target_size=(size_t)(nelmts*max_type_size);
     } /* end if */
     request_nelmts = target_size / max_type_size;
         
@@ -1724,7 +1724,7 @@ H5D_chunk_read(hsize_t nelmts, H5D_t *dataset,
         for (smine_start=0; smine_start<chunk_info->chunk_points; smine_start+=smine_nelmts) {
             /* Go figure out how many elements to read from the file */
             assert(H5S_SELECT_ITER_NELMTS(&file_iter)==(chunk_info->chunk_points-smine_start));
-            smine_nelmts = MIN(request_nelmts, (chunk_info->chunk_points-smine_start));
+            smine_nelmts = (size_t)MIN(request_nelmts, (chunk_info->chunk_points-smine_start));
         
             /*
              * Gather the data from disk into the data type conversion
@@ -2004,7 +2004,7 @@ H5D_chunk_write(hsize_t nelmts, H5D_t *dataset,
             target_size = max_type_size;
         /* If the buffer is too large to hold all the elements, make it smaller */
         else if(target_size>(nelmts*max_type_size))
-            target_size=(nelmts*max_type_size);
+            target_size=(size_t)(nelmts*max_type_size);
     } /* end if */
     request_nelmts = target_size / max_type_size;
 
@@ -2074,7 +2074,7 @@ H5D_chunk_write(hsize_t nelmts, H5D_t *dataset,
         for (smine_start=0; smine_start<chunk_info->chunk_points; smine_start+=smine_nelmts) {
             /* Go figure out how many elements to read from the file */
             assert(H5S_SELECT_ITER_NELMTS(&file_iter)==(chunk_info->chunk_points-smine_start));
-            smine_nelmts = MIN(request_nelmts, (chunk_info->chunk_points-smine_start));
+            smine_nelmts = (size_t)MIN(request_nelmts, (chunk_info->chunk_points-smine_start));
             
             /*
              * Gather the data from disk into the data type conversion
